+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Usage section to README.
+- `junk-setup-use-package` to make sure ensuring does not install
+  `junk` pack packages.
 
 ### Changed